moved brachy applicator to geometry utils
